requirements.txt
   openai
   openai[datalib]
   python-dotenv
   grpcio
   grpcio-tools
   grpcio-testing
   protobuf